+++ app/components/ember-youtube.js
@@ -1,4 +1,3 @@
-import Ember from 'ember';
 import EmberYoutube from 'ember-youtube/components/ember-youtube';
 
 export default EmberYoutube;