+++ tests/dummy/app/controllers/application.js
@@ -1,7 +1,8 @@
 import Ember from 'ember';
 
 export default Ember.Controller.extend({
-	youTubeId: 'fZ7MhTRmJ60', // NEFrNP-BLcI
+	youTubeId: 'fZ7MhTRmJ60',
+	volume: 100,
 
 	customPlayerVars: {
 		autoplay: 1,
@@ -9,7 +10,6 @@ export default Ember.Controller.extend({
 		showinfo: 0 // hide uploader info
 	},
 
-	volume: 100,
 	actions: {
 		ytPlaying() {
 			Ember.debug('on playing from controller');